fixed the composable function in the main feature

+++ news-data/src/main/java/com/verinskij/news/data/AllArticlesPagingSource.kt
@@ -43,6 +43,6 @@ class AllArticlesPagingSource(
     }
 
     private companion object {
-        var counter: Long = 1
+        var counter: Int = 1
     }
 }
